Refactor: Move Time Implementation out of Logic
* Refactor: Move Time Setup out of Logic
  * Logic Time: Extract BusTime Setup
  * Common Time: Move Timer Setup out of Logic
* Common Time: Move Time/Date-Read-Requests out of Logic
* Common Time: Cleanup busTime_setup
* Common Time: Move `busTime_processInputKo` out of Logic + Use DPT Definitions
