+++ src/Timer.cpp
@@ -1,6 +1,5 @@
 #include "Timer.h"
 #include "Arduino.h"
-#include "OpenKNX.h"
 #include <ctime>
 
 #ifdef OPENKNX_EXPERIMENTAL_RP2040RTC_LOCALTIME
@@ -69,12 +68,16 @@ Timer &Timer::instance()
     return sInstance;
 }
 
-void Timer::setup(double iLongitude, double iLatitude, int8_t iTimezone, bool iUseSummertime, uint64_t iHolidayBitmask)
+void Timer::setup(uint64_t iHolidayBitmask)
 {
+    bool lTimezoneSign = ParamBASE_TimezoneSign;
+    int8_t lTimezone = ParamBASE_TimezoneValue;
+    lTimezone = lTimezone * (lTimezoneSign ? -1 : 1);
+    bool iUseSummertime = (ParamBASE_SummertimeAll == VAL_STIM_FROM_INTERN);
 
-    mLongitude = iLongitude;
-    mLatitude = iLatitude;
-    mTimezone = iTimezone;
+    mLongitude = ParamBASE_Longitude;
+    mLatitude = ParamBASE_Latitude;
+    mTimezone = ParamBASE_Timezone;
     mUseSummertime = iUseSummertime;
     // we delete all unnecessary holidays from holiday data
     for (uint8_t i = 0; i < cHolidaysCount; i++)
@@ -90,8 +93,56 @@ bool Timer::UseSummertime()
     return mUseSummertime;
 }
 
+// TODO Move to BusTime
+void Timer::busTime_processReadRequests()
+{
+    static uint32_t sDelay = 19000;
+
+    // date and time are red from bus every 30 seconds until a response is received
+    if (ParamBASE_ReadTimeDate)
+    {
+        const eTimeValid lValid = isTimerValid();
+        if (delayCheck(sDelay, 30000) && lValid != tmValid)
+        {
+            // logDebugP("Time Valid? %i", lValid);
+            sDelay = millis();
+            if (ParamBASE_CombinedTimeDate)
+            {
+                // combined date and time
+                KoBASE_Time.requestObjectRead();
+            }
+            else
+            {
+                // date and time from separate KOs
+                if (lValid != tmMinutesValid)
+                    KoBASE_Time.requestObjectRead();
+                if (lValid != tmDateValid)
+                    KoBASE_Date.requestObjectRead();
+            }
+        }
+        // if date and/or time is known, we read also summertime information
+        // TODO check dependency to configuration
+        if (sDelay > 0 && lValid == tmValid)
+        {
+            sDelay = 0;
+            KoBASE_IsSummertime.requestObjectRead();
+        }
+    }
+}
+
+void Timer::busTime_loop()
+{
+    // TODO needed after Separation only
+    if (!openknx.afterStartupDelay())
+        return;
+
+    busTime_processReadRequests();
+}
+
 void Timer::loop()
 {
+    busTime_loop();
+
     if (mTimeDelay == 0 || delayCheck(mTimeDelay, 1000))
     {
         // if time is set from bus, we have immediately to recalculate everything which is necessary
@@ -144,6 +195,79 @@ void Timer::loop()
                 mDayTick = mNow.tm_mday;
             }
         }
+    }
+}
+
+void Timer::busTime_processInputKo(GroupObject &iKo)
+{
+    if (iKo.asap() == BASE_KoTime)
+    {
+        if (ParamBASE_CombinedTimeDate)
+        {
+            KNXValue value = "";
+
+            // first ensure we have a valid data-time content
+            // (including the correct length)
+            if (iKo.tryValue(value, DPT_DateTime))
+            {
+
+                // use raw value, as current version of knx do not provide access to all fields
+                // TODO DPT19: check integration of extended DPT19 access into knx or OpenKNX-Commons
+                // size is ensured to be 8 Byte
+                uint8_t *raw = iKo.valueRef();
+
+                /*
+                const bool flagFault = raw[6] & 0x80;
+                // ignore working day (WD, NWD): raw[6] & 0x40, raw[6] & 0x20
+                const bool flagNoYear = raw[6] & 0x10;
+                const bool flagNoDate = raw[6] & 0x08;
+                // ignore NDOW: raw[6] & 0x04
+                const bool flagNoTime = raw[6] & 0x02;
+                const bool flagSuti = raw[6] & 0x01;
+                // ignore quality of clock (CLQ): raw[7] & 0x80
+                // ignore synchronisation source reliablity (SRC): raw[7] & 0x40
+                */
+
+                // ignore inputs with:
+                // * F - fault
+                // * NY - missing year
+                // * ND - missing date
+                // * NT - missing time
+                if (!(raw[6] & (DPT19_FAULT | DPT19_NO_YEAR | DPT19_NO_DATE | DPT19_NO_TIME)))
+                {
+                    struct tm lTmp = value;
+                    setDateTimeFromBus(&lTmp);
+                    const bool lSummertime = raw[6] & DPT19_SUMMERTIME;
+                    // TODO check using ParamLOG_SummertimeAll
+                    if (((knx.paramByte(BASE_SummertimeAll) & BASE_SummertimeAllMask) >> BASE_SummertimeAllShift) == VAL_STIM_FROM_DPT19)
+                        IsSummertime(lSummertime);
+                }
+            }
+        }
+        else
+        {
+            KNXValue value = "";
+            // ensure we have a valid time content
+            if (iKo.tryValue(value, DPT_TimeOfDay))
+            {
+                struct tm lTmp = value;
+                setTimeFromBus(&lTmp);
+            }
+        }
+    }
+    else if (iKo.asap() == BASE_KoDate)
+    {
+        KNXValue value = "";
+        // ensure we have a valid date content
+        if (iKo.tryValue(value, DPT_Date))
+        {
+            struct tm lTmp = value;
+            setDateFromBus(&lTmp);
+        }
+    }
+    else if (iKo.asap() == BASE_KoIsSummertime)
+    {
+        IsSummertime(iKo.value(DPT_Date));
     }
 }
